feat: token cleanup and session token middleware + tests about it

+++ source/src/auth/AuthenticationRouter.js
@@ -11,7 +11,6 @@ const router = express.Router();
 
 router.post('/api/1.0/auth', loginValidation, async (req, res, next) => {
   const { email, password } = req.body;
-
   const user = await UserService.getUserByEmail(email);
   if (!user) return next(new AuthenticationException()); // 401
 
@@ -20,19 +19,21 @@ router.post('/api/1.0/auth', loginValidation, async (req, res, next) => {
 
   if (user.inactive) return next(new ForbiddenException()); // 403
   const token = await TokenService.createToken({ id: user.id });
+
   return res.status(200).send({ id: user.id, username: user.username, token });
 });
 
 router.post('/api/1.0/logout/:id', tokenAuthentication, async (req, res, next) => {
-  console.log('AUTHN USER', req.authenticatedUser);
+  // if no auth header is sent, then 204 ok no content is sent. (debatable, le pregunté a Ale y así lo hace)
+  if (!req.headers.authorization) {
+    return res.status(204).send();
+  }
   if (!req.authenticatedUser || req.authenticatedUser.id != req.params.id) {
-    console.log('entra al forbidden');
-    return next(new ForbiddenException(req.t('unauthorized_user_loggout')));
+    return next(new ForbiddenException(req.t('unauthorized_user_logout')));
   }
   const token = req.headers.authorization.substring(7);
-  console.log('tooken', token);
   await TokenService.deleteToken(token);
-  return res.status(200).send();
+  return res.status(204).send();
 });
 
 module.exports = router;
